added abstract class for visitor
added classes for NumberConst, UnaryOperation, BinaryOperation

+++ include/evaluable/evaluable.hpp
@@ -6,6 +6,6 @@
 class Evaluable : public AstNode
 {
 public:
-    virtual int acceptVisitor(Visitor *) = 0;
+    virtual double acceptVisitor(Visitor *) = 0;
     virtual ~Evaluable() = 0;
 };
